Allow piecemeal overrides of tokenserver backend default config.
Before this change, you have either have to accept the default config
for the tokenserver backend, or specify all config options and hence
duplicate info from earlier in the file.  With this change you can
now just specify anything you want to change and we'll use the defaults
for the rest.